+++ js/util.js
@@ -1,0 +1,5 @@
+function getRandomInt(min, max) {
+  return min + Math.random() * (max - min + 1) >> 0;
+}
+
+export {getRandomInt};
